+++ Source/Gpu.hpp
@@ -15,7 +15,7 @@ struct Gpu
 		OAM = 0x2, Transfer = 0x3
 	};
 
-	uint16_t clock;
+	int32_t clock;
 	
 	union {
 		uint8_t value;
